requirements/production.txt
-r base.txt
# elasticsearch==2.x.x chosen for compatibility with t2.micro.elasticsearch and t2.small.elasticsearch
# instance types on AWS (Elasticsearch 2.3). Adjust for your deployment as needed.
elasticsearch==6.8.1
# Additional dependencies for Heroku, AWS, and Google Cloud deployment
dj-database-url==0.4.1
uwsgi>=2.0.17,<2.1
psycopg2>=2.7,<3.0
whitenoise>=5.0,<5.1
boto3==1.9.189
google-cloud-storage==1.20.0
django-storages==1.11.1
# For retrieving credentials and signing requests to Elasticsearch
botocore
aws-requests-auth==0.4.0
django-redis==4.11.0
django_cache_url==2.0.0
wagtail-bakery